+++ app/src/organisms/Devices/hooks/index.ts
@@ -23,3 +23,4 @@ export * from './useProtocolAnalysisErrors'
 export * from './useRunStartedOrLegacySessionInProgress'
 export * from './useProtocolRunAnalyticsData'
 export * from './useRobotAnalyticsData'
+export * from './useTrackProtocolRunEvent'
